+++ vrpy/checks.py
@@ -8,15 +8,17 @@ from networkx import DiGraph, NetworkXError, has_path
 logger = logging.getLogger(__name__)
 
 
-def check_arguments(num_stops: int = None,
-                    load_capacity: list = None,
-                    duration: int = None,
-                    pricing_strategy: str = None,
-                    mixed_fleet: bool = None,
-                    fixed_cost: bool = None,
-                    G: DiGraph = None,
-                    vehicle_types: int = None,
-                    num_vehicles: list = None):
+def check_arguments(
+    num_stops: int = None,
+    load_capacity: list = None,
+    duration: int = None,
+    pricing_strategy: str = None,
+    mixed_fleet: bool = None,
+    fixed_cost: bool = None,
+    G: DiGraph = None,
+    vehicle_types: int = None,
+    num_vehicles: list = None,
+):
     """Checks if arguments are consistent."""
 
     # If num_stops/load_capacity/duration are not integers
@@ -30,32 +32,33 @@ def check_arguments(num_stops: int = None,
         raise TypeError("Maximum duration must be positive integer.")
     strategies = ["Exact", "BestEdges1", "BestEdges2", "BestPaths", "Hyper"]
     if pricing_strategy not in strategies:
-        raise ValueError("Pricing strategy %s is not valid. Pick one among %s" %
-                         (pricing_strategy, strategies))
+        raise ValueError(
+            "Pricing strategy %s is not valid. Pick one among %s"
+            % (pricing_strategy, strategies)
+        )
     if mixed_fleet:
-        if (load_capacity and num_vehicles and
-                len(load_capacity) != len(num_vehicles)):
+        if load_capacity and num_vehicles and len(load_capacity) != len(num_vehicles):
             raise ValueError(
                 "Input arguments load_capacity and num_vehicles must have same dimension."
             )
-        if (load_capacity and fixed_cost and
-                len(load_capacity) != len(fixed_cost)):
+        if load_capacity and fixed_cost and len(load_capacity) != len(fixed_cost):
             raise ValueError(
                 "Input arguments load_capacity and fixed_cost must have same dimension."
             )
-        if (num_vehicles and fixed_cost and
-                len(num_vehicles) != len(fixed_cost)):
+        if num_vehicles and fixed_cost and len(num_vehicles) != len(fixed_cost):
             raise ValueError(
                 "Input arguments num_vehicles and fixed_cost must have same dimension."
             )
         for (i, j) in G.edges():
             if not isinstance(G.edges[i, j]["cost"], list):
                 raise TypeError(
-                    "Cost attribute for edge (%s,%s) should be of type list")
+                    "Cost attribute for edge (%s,%s) should be of type list"
+                )
             if len(G.edges[i, j]["cost"]) != vehicle_types:
                 raise ValueError(
                     "Cost attribute for edge (%s,%s) has dimension %s, should have dimension %s."
-                    % (i, j, len(G.edges[i, j]["cost"]), vehicle_types))
+                    % (i, j, len(G.edges[i, j]["cost"]), vehicle_types)
+                )
 
 
 def check_vrp(G: DiGraph = None):
@@ -63,8 +66,7 @@ def check_vrp(G: DiGraph = None):
 
     # if G is not a DiGraph
     if not isinstance(G, DiGraph):
-        raise TypeError(
-            "Input graph must be of type networkx.classes.digraph.DiGraph.")
+        raise TypeError("Input graph must be of type networkx.classes.digraph.DiGraph.")
     for v in ["Source", "Sink"]:
         # If Source or Sink is missing
         if v not in G.nodes():
@@ -105,8 +107,7 @@ def check_initial_routes(initial_routes: list = None, G: DiGraph = None):
 
     for route in initial_routes:
         if route[0] != "Source" or route[-1] != "Sink":
-            raise ValueError("Route %s must start at Source and end at Sink" %
-                             route)
+            raise ValueError("Route %s must start at Source and end at Sink" % route)
     # Check if every node is in at least one route
     for v in G.nodes():
         if v not in ["Source", "Sink"]:
@@ -121,16 +122,19 @@ def check_initial_routes(initial_routes: list = None, G: DiGraph = None):
         edges = list(zip(route[:-1], route[1:]))
         for (i, j) in edges:
             if (i, j) not in G.edges():
-                raise KeyError("Edge (%s,%s) in route %s missing in graph." %
-                               (i, j, route))
+                raise KeyError(
+                    "Edge (%s,%s) in route %s missing in graph." % (i, j, route)
+                )
             if "cost" not in G.edges[i, j]:
                 raise KeyError("Edge (%s,%s) has no cost attribute." % (i, j))
 
 
-def check_consistency(cspy: bool = None,
-                      pickup_delivery: bool = None,
-                      pricing_strategy: str = None,
-                      G: DiGraph = None):
+def check_consistency(
+    cspy: bool = None,
+    pickup_delivery: bool = None,
+    pricing_strategy: str = None,
+    G: DiGraph = None,
+):
     """Raises errors if options are inconsistent with parameters."""
 
     # pickup delivery requires cspy=False
@@ -144,31 +148,34 @@ def check_consistency(cspy: bool = None,
     if pickup_delivery:
         request = any("request" in G.nodes[v] for v in G.nodes())
         if not request:
-            raise KeyError(
-                "pickup_delivery option expects at least one request.")
+            raise KeyError("pickup_delivery option expects at least one request.")
 
 
-def check_feasibility(load_capacity: list = None,
-                      G: DiGraph = None,
-                      duration: int = None):
+def check_feasibility(
+    load_capacity: list = None, G: DiGraph = None, duration: int = None
+):
     """Checks basic problem feasibility."""
 
     if load_capacity:
         for v in G.nodes():
             if G.nodes[v]["demand"] > max(load_capacity):
                 raise ValueError(
-                    "Demand %s at node %s larger than max capacity %s." %
-                    (G.nodes[v]["demand"], v, max(load_capacity)))
+                    "Demand %s at node %s larger than max capacity %s."
+                    % (G.nodes[v]["demand"], v, max(load_capacity))
+                )
     if duration:
         for v in G.nodes():
             if v not in ["Source", "Sink"]:
-                round_trip_duration = (G.nodes[v]["service_time"] +
-                                       G.edges["Source", v]["time"] +
-                                       G.edges[v, "Sink"]["time"])
+                round_trip_duration = (
+                    G.nodes[v]["service_time"]
+                    + G.edges["Source", v]["time"]
+                    + G.edges[v, "Sink"]["time"]
+                )
                 if round_trip_duration > duration:
                     raise ValueError(
                         "Node %s not reachable: duration of path [Source,%s,Sink], %s, is larger than max duration %s."
-                        % (v, v, round_trip_duration, duration))
+                        % (v, v, round_trip_duration, duration)
+                    )
 
 
 def check_seed(seed):
@@ -183,3 +190,30 @@ def check_seed(seed):
         return seed
     else:
         raise TypeError("{} cannot be used to seed".format(seed))
+
+
+def check_pickup_delivery_time_windows(G: DiGraph, edges: list):
+    for (i, j) in edges:
+        if (i, j) not in G.edges():
+            raise ValueError(
+                "Problem Infeasible, request (%s,%s) cannot be done with given time windows."
+                % (i, j)
+            )
+
+
+def check_periodic_num_vehicles(periodic=None, num_vehicles=[]):
+    """
+    The case where periodic is not None and num_vehicles is not None is ambiguous
+    and needs to be more specifically defined.
+    For the moment, if both are activated, num_vehicles is ignored until the final
+    schedule is computed (vrpy/schedule.py)
+    """
+    _num_vehicles = []
+    _num_vehicles_schedule = []
+    if periodic and num_vehicles:
+        # if both are activated, _num_vehicles is None and _num_vehicles_schedule = num_vehicles
+        _num_vehicles_schedule = num_vehicles
+    else:
+        # if either is not active, _num_vehicles = num_vehicles, _num_vehicles_schedule is None
+        _num_vehicles = num_vehicles
+    return _num_vehicles, _num_vehicles_schedule