+++ vrpy/subproblem_cspy.py
@@ -1,11 +1,8 @@
 import logging
 from math import floor
-
-from numpy import array, zeros
+from numpy import zeros
 from networkx import DiGraph, add_path
-
 from cspy import BiDirectional, REFCallback
-
 from vrpy.subproblem import _SubProblemBase
 
 logger = logging.getLogger(__name__)
@@ -17,33 +14,22 @@ class _MyREFCallback(REFCallback):
     Based on Righini and Salani (2006).
     """
 
-    def __init__(self, max_res, sub_G, time_windows, distribution_collection, T,
-                 resources):
+    def __init__(self, max_res, time_windows, distribution_collection, T, resources):
         REFCallback.__init__(self)
         # Set attributes for use in REF functions
         self._max_res = max_res
-        self._sub_G = sub_G
         self._time_windows = time_windows
         self._distribution_collection = distribution_collection
         self._T = T
         self._resources = resources
-        self._original_node_type = type(
-            [n for n in self._sub_G.nodes() if n not in ["Source", "Sink"]][0])
+        # Set later
+        self._sub_G = None
+        self._source_id = None
+        self._sink_id = None
 
-    def REF_fwd(self, cumul_res, tail, head, edge_res, partial_path,
-                cumul_cost):
+    def REF_fwd(self, cumul_res, tail, head, edge_res, partial_path, cumul_cost):
         new_res = list(cumul_res)
-
-        if tail != "Source":
-            i = self._convert_to_original_type(tail)
-        else:
-            i = tail
-
-        if head != "Sink":
-            j = self._convert_to_original_type(head)
-        else:
-            j = head
-
+        i, j = tail, head
         # stops / monotone resource
         new_res[0] += 1
         # load
@@ -51,21 +37,21 @@ class _MyREFCallback(REFCallback):
         # time
         # Service times
         theta_i = self._sub_G.nodes[i]["service_time"]
-        theta_j = self._sub_G.nodes[j]["service_time"]
-        theta_t = self._sub_G.nodes["Sink"]["service_time"]
+        # theta_j = self._sub_G.nodes[j]["service_time"]
+        # theta_t = self._sub_G.nodes[self._sink_id]["service_time"]
         # Travel times
         travel_time_ij = self._sub_G.edges[i, j]["time"]
-        try:
-            travel_time_jt = self._sub_G.edges[j, "Sink"]["time"]
-        except KeyError:
-            travel_time_jt = 0
+        # try:
+        #    travel_time_jt = self._sub_G.edges[j, self._sink_id]["time"]
+        # except KeyError:
+        #    travel_time_jt = 0
         # Time windows
         # Lower
         a_j = self._sub_G.nodes[j]["lower"]
-        a_t = self._sub_G.nodes["Sink"]["lower"]
+        # a_t = self._sub_G.nodes[self._sink_id]["lower"]
         # Upper
         b_j = self._sub_G.nodes[j]["upper"]
-        b_t = self._sub_G.nodes["Sink"]["upper"]
+        # b_t = self._sub_G.nodes[self._sink_id]["upper"]
 
         new_res[2] = max(new_res[2] + theta_i + travel_time_ij, a_j)
 
@@ -73,6 +59,9 @@ class _MyREFCallback(REFCallback):
         if not self._time_windows or (new_res[2] <= b_j):
             # and new_res[2] < self._T - a_j - theta_j and
             # a_t <= new_res[2] + travel_time_jt + theta_t <= b_t):
+            # if not self._time_windows or (
+            #         new_res[2] <= b_j and new_res[2] < self._T - a_j - theta_j and
+            #         a_t <= new_res[2] + travel_time_jt + theta_t <= b_t):
             new_res[3] = 0
         else:
             new_res[3] = 1
@@ -81,116 +70,92 @@ class _MyREFCallback(REFCallback):
             # Pickup
             new_res[4] += self._sub_G.nodes[j]["collect"]
             # Delivery
-            new_res[5] = max(new_res[5] + self._sub_G.nodes[j]["demand"],
-                             new_res[4])
+            new_res[5] = max(new_res[5] + self._sub_G.nodes[j]["demand"], new_res[4])
         return new_res
 
-    # def REF_bwd(self, cumul_res, tail, head, edge_res, partial_path,
-    #             cumul_cost):
-    #     new_res = list(cumul_res)
+    def REF_bwd(self, cumul_res, tail, head, edge_res, partial_path, cumul_cost):
+        new_res = list(cumul_res)
+        i, j = tail, head
 
-    #     if tail != "Source":
-    #         i = self._convert_to_original_type(tail)
-    #     else:
-    #         i = tail
+        # monotone resource
+        new_res[0] -= 1
+        # load
+        new_res[1] += self._sub_G.nodes[i]["demand"]
+        # Get relevant service times (thetas) and travel time
+        # Service times
+        theta_i = self._sub_G.nodes[i]["service_time"]
+        theta_j = self._sub_G.nodes[j]["service_time"]
+        # theta_s = self._sub_G.nodes[self._source_id]["service_time"]
+        # Travel times
+        travel_time_ij = self._sub_G.edges[i, j]["time"]
+        # try:
+        #    travel_time_si = self._sub_G.edges[self._source_id, i]["time"]
+        # except KeyError:
+        #    travel_time_si = 0
+        # Lower time windows
+        # a_i = self._sub_G.nodes[i]["lower"]
+        # a_j = self._sub_G.nodes[j]["lower"]
+        # a_s = self._sub_G.nodes[self._source_id]["lower"]
+        # Upper time windows
+        b_i = self._sub_G.nodes[i]["upper"]
+        b_j = self._sub_G.nodes[j]["upper"]
+        # b_s = self._sub_G.nodes[self._source_id]["upper"]
 
-    #     if head != "Sink":
-    #         j = self._convert_to_original_type(head)
-    #     else:
-    #         j = head
+        new_res[2] = max(new_res[2] + theta_j + travel_time_ij, self._T - b_i - theta_i)
 
-    #     # self._convert_to_original_type(tail, head)
-    #     # monotone resource
-    #     new_res[0] -= 1
-    #     # load
-    #     new_res[1] += self._sub_G.nodes[i]["demand"]
-    #     # Get relevant service times (thetas) and travel time
-    #     # Service times
-    #     theta_i = self._sub_G.nodes[i]["service_time"]
-    #     theta_j = self._sub_G.nodes[j]["service_time"]
-    #     theta_s = self._sub_G.nodes["Source"]["service_time"]
-    #     # Travel times
-    #     travel_time_ij = self._sub_G.edges[i, j]["time"]
-    #     try:
-    #         travel_time_si = self._sub_G.edges["Source", i]["time"]
-    #     except KeyError:
-    #         travel_time_si = 0
-    #     # Lower time windows
-    #     a_i = self._sub_G.nodes[i]["lower"]
-    #     a_s = self._sub_G.nodes["Source"]["lower"]
-    #     # Upper time windows
-    #     b_i = self._sub_G.nodes[i]["upper"]
-    #     b_j = self._sub_G.nodes[j]["upper"]
-    #     b_s = self._sub_G.nodes["Source"]["upper"]
+        # time-window feasibility
+        if not self._time_windows or (new_res[2] <= b_j):
+            # and new_res[2] < self._T - a_i - theta_i and
+            #         a_s <= new_res[2] + theta_s + travel_time_si <= b_s):
+            # if not self._time_windows or (
+            #         new_res[2] <= self._T - a_j and
+            #         new_res[2] < self._T - a_i - theta_i and
+            #         a_s <= new_res[2] + theta_s + travel_time_si <= b_s):
+            new_res[3] = 0
+        else:
+            new_res[3] = 1
 
-    #     new_res[2] = max(new_res[2] + theta_j + travel_time_ij,
-    #                      self._T - b_i - theta_i)
+        if self._distribution_collection:
+            # Delivery
+            new_res[5] += new_res[5] + self._sub_G.nodes[i]["demand"]
+            # Pickup
+            new_res[4] = max(new_res[5], new_res[4] + self._sub_G.nodes[i]["collect"])
+        return new_res
 
-    #     # time-window feasibility
-    #     if not self._time_windows or (new_res[2] <= b_j):
-    #         # and new_res[2] < self._T - a_i - theta_i and
-    #         #         a_s <= new_res[2] + theta_s + travel_time_si <= b_s):
-    #         new_res[3] = 0
-    #     else:
-    #         new_res[3] = 1
+    def REF_join(self, fwd_resources, bwd_resources, tail, head, edge_res):
+        """
+        Appropriate joining of forward and backward resources.
+        """
+        fwd_res = list(fwd_resources)
+        bwd_res = list(bwd_resources)
+        final_res = [0] * len(self._resources)
 
-    #     if self._distribution_collection:
-    #         # Delivery
-    #         new_res[5] += new_res[5] + self._sub_G.nodes[i]["demand"]
-    #         # Pickup
-    #         new_res[4] = max(new_res[5],
-    #                          new_res[4] + self._sub_G.nodes[i]["collect"])
-    #     return new_res
+        i, j = tail, head
 
-    # def REF_join(self, fwd_resources, bwd_resources, tail, head, edge_res):
-    #     """
-    #     Appropriate joining of forward and backward resources.
-    #     """
-    #     fwd_res = list(fwd_resources)
-    #     bwd_res = list(bwd_resources)
-    #     final_res = [0] * len(self._resources)
+        # Get relevant service times (thetas) and travel time
+        theta_i = self._sub_G.nodes[i]["service_time"]
+        theta_j = self._sub_G.nodes[j]["service_time"]
+        travel_time = self._sub_G.edges[i, j]["time"]
 
-    #     if tail != "Source":
-    #         i = self._convert_to_original_type(tail)
-    #     else:
-    #         i = tail
+        # Invert monotone resource
+        bwd_res[0] = self._max_res[0] - bwd_res[0]
+        # Fill in final res
+        # Monotone / stops
+        final_res[0] = fwd_res[0] + bwd_res[0] + 1
+        # Load
+        final_res[1] = fwd_res[1] + bwd_res[1]
+        # time
+        final_res[2] = fwd_res[2] + theta_i + travel_time + theta_j + bwd_res[2]
+        # Time windows
+        if not self._time_windows or final_res[2] <= self._T:
+            final_res[3] = fwd_res[3] + bwd_res[3]
+        else:
+            final_res[3] = 1.0
 
-    #     if head != "Sink":
-    #         j = self._convert_to_original_type(head)
-    #     else:
-    #         j = head
-
-    #     # Get relevant service times (thetas) and travel time
-    #     theta_i = self._sub_G.nodes[i]["service_time"]
-    #     theta_j = self._sub_G.nodes[j]["service_time"]
-    #     travel_time = self._sub_G.edges[i, j]["time"]
-
-    #     # Invert monotone resource
-    #     bwd_res[0] = self._max_res[0] - bwd_res[0]
-    #     # Fill in final res
-    #     # Monotone / stops
-    #     final_res[0] = fwd_res[0] + bwd_res[0] + 1
-    #     # Load
-    #     final_res[1] = fwd_res[1] + bwd_res[1]
-    #     # time
-    #     final_res[2] = fwd_res[2] + theta_i + travel_time + theta_j + bwd_res[2]
-    #     # Time windows
-    #     if not self._time_windows or final_res[2] <= self._T:
-    #         final_res[3] = fwd_res[3] + bwd_res[3]
-    #     else:
-    #         final_res[3] = 1.0
-
-    #     if self._distribution_collection:
-    #         final_res[4] = fwd_res[4] + bwd_res[4]
-    #         final_res[5] = fwd_res[5] + bwd_res[5]
-    #     return final_res
-
-    def _convert_to_original_type(self, node):
-        if "int" in self._original_node_type.__name__:
-            n = int(node)
-        elif "str" in self._original_node_type.__name__:
-            n = str(node)
-        return n
+        if self._distribution_collection:
+            final_res[4] = fwd_res[4] + bwd_res[4]
+            final_res[5] = fwd_res[5] + bwd_res[5]
+        return final_res
 
 
 class _SubProblemCSPY(_SubProblemBase):
@@ -220,15 +185,13 @@ class _SubProblemCSPY(_SubProblemBase):
         # Default lower and upper bounds
         self.min_res = [0] * len(self.resources)
         # Add upper bounds for mono, stops, load and time, and time windows
-        total_demand = sum(
-            [self.sub_G.nodes[v]["demand"] for v in self.sub_G.nodes()])
+        total_demand = sum([self.sub_G.nodes[v]["demand"] for v in self.sub_G.nodes()])
         self.max_res = [
-            len(self.sub_G.nodes()),  # stop/mono
-            # floor(len(self.sub_G.nodes()) / 2),  # stop/mono
+            floor(len(self.sub_G.nodes()) / 2),  # stop/mono
             total_demand,  # load
-            # floor(total_demand / 2),  # load
-            sum([self.sub_G.edges[u, v]["time"] for u, v in self.sub_G.edges()]
-               ),  # time
+            sum(
+                [self.sub_G.edges[u, v]["time"] for u, v in self.sub_G.edges()]
+            ),  # time
             1,  # time windows
             total_demand,  # pickup
             total_demand,  # deliver
@@ -261,31 +224,39 @@ class _SubProblemCSPY(_SubProblemBase):
 
         more_routes = False
 
+        my_callback = self.get_REF()
         while True:
             if self.exact:
-                alg = BiDirectional(self.sub_G,
-                                    self.max_res,
-                                    self.min_res,
-                                    direction="forward",
-                                    time_limit=time_limit -
-                                    0.5 if time_limit else None,
-                                    elementary=True,
-                                    REF_callback=self.get_REF())
+                alg = BiDirectional(
+                    self.sub_G,
+                    self.max_res,
+                    self.min_res,
+                    direction="both",
+                    time_limit=time_limit - 0.5 if time_limit else None,
+                    elementary=True,
+                    REF_callback=my_callback,
+                )
             else:
-                alg = BiDirectional(self.sub_G,
-                                    self.max_res,
-                                    self.min_res,
-                                    threshold=-1,
-                                    direction="forward",
-                                    time_limit=time_limit -
-                                    0.5 if time_limit else None,
-                                    elementary=True,
-                                    REF_callback=self.get_REF())
+                alg = BiDirectional(
+                    self.sub_G,
+                    self.max_res,
+                    self.min_res,
+                    threshold=-1,
+                    direction="both",
+                    time_limit=time_limit - 0.5 if time_limit else None,
+                    elementary=True,
+                    REF_callback=my_callback,
+                )
+
+            # Pass processed graph
+            if my_callback is not None:
+                my_callback._sub_G = alg.G
+                my_callback._source_id = alg._source_id
+                my_callback._sink_id = alg._sink_id
             alg.run()
             logger.debug("subproblem")
             logger.debug("cost = %s", alg.total_cost)
             logger.debug("resources = %s", alg.consumed_resources)
-            print(alg.path)
             if alg.total_cost is not None and alg.total_cost < -(1e-3):
                 more_routes = True
                 self.add_new_route(alg.path)
@@ -321,10 +292,12 @@ class _SubProblemCSPY(_SubProblemBase):
             # Time windows feasibility
             self.max_res[3] = 0
             # Maximum feasible arrival time
-            self.T = max(self.sub_G.nodes[v]["upper"] +
-                         self.sub_G.nodes[v]["service_time"] +
-                         self.sub_G.edges[v, "Sink"]["time"]
-                         for v in self.sub_G.predecessors("Sink"))
+            self.T = max(
+                self.sub_G.nodes[v]["upper"]
+                + self.sub_G.nodes[v]["service_time"]
+                + self.sub_G.edges[v, "Sink"]["time"]
+                for v in self.sub_G.predecessors("Sink")
+            )
 
         if self.load_capacity and self.distribution_collection:
             self.max_res[4] = self.load_capacity[self.vehicle_type]
@@ -379,9 +352,13 @@ class _SubProblemCSPY(_SubProblemBase):
     def get_REF(self):
         if self.time_windows or self.distribution_collection:
             # Use custom REF
-            return _MyREFCallback(self.max_res, self.sub_G, self.time_windows,
-                                  self.distribution_collection, self.T,
-                                  self.resources)
+            return _MyREFCallback(
+                self.max_res,
+                self.time_windows,
+                self.distribution_collection,
+                self.T,
+                self.resources,
+            )
         else:
             # Use default
-            return
+            return None